+++ client/components/itemDetail/itemDetailForm.spec.js
@@ -10,11 +10,52 @@ import ItemDetailForm from './ItemDetailForm.jsx';
 import { expect } from 'chai';
 
 describe('<ItemDetailForm />', () => {
+
   const errors = {
     name: false,
     label: false,
     price: false
   };
+
+  const categories = [
+    {
+      categoryID: 1,
+      categoryName: 'Foo',
+      companyID: 1,
+      createdBy: 1,
+      createdDate: '2017-03-31T01:09:34.3905613-07:00',
+      facilityID: 0,
+      isActive: true,
+      items: [],
+      lastUpdatedBy: 1,
+      lastUpdatedDate: '2017-03-31T01:09:34.3905613-07:00'
+    },
+    {
+      categoryID: 2,
+      categoryName: 'Qux',
+      companyID: 1,
+      createdBy: 1,
+      createdDate: '2017-03-31T01:09:34.3905613-07:00',
+      facilityID: 0,
+      isActive: true,
+      items: [],
+      lastUpdatedBy: 1,
+      lastUpdatedDate: '2017-03-31T01:09:34.3905613-07:00'
+    },
+    {
+      categoryID: 3,
+      categoryName: 'Bar',
+      companyID: 1,
+      createdBy: 1,
+      createdDate: '2017-03-31T01:09:34.3905613-07:00',
+      facilityID: 0,
+      isActive: true,
+      items: [],
+      lastUpdatedBy: 1,
+      lastUpdatedDate: '2017-03-31T01:09:34.3905613-07:00'
+    }
+  ];
+
   const props = {
     item: {
       itemID: 0,
@@ -28,6 +69,7 @@ describe('<ItemDetailForm />', () => {
       isActive: 1,
       priceTypeID: 0
     },
+    categories,
     errors,
     onChange() { },
     onDrop() { }
@@ -56,23 +98,36 @@ describe('<ItemDetailForm />', () => {
 
   });
 
-  it('should render a SelectList component', () => {
+  it('should render two SelectList components', () => {
 
     const wrapper = shallow(<ItemDetailForm {...props} />);
 
-    const expected = 1;
+    const expected = 2;
     const actual = wrapper.find(SelectList).length;
 
     expect(actual).equals(expected);
 
   });
 
-  it('should render 4 list items on the SelectListComponent', () => {
+  it('should render 4 list items on the SelectListComponent for priceType', () => {
 
     const wrapper = shallow(<ItemDetailForm {...props} />);
 
     const expected = 4;
-    const actual = wrapper.find(SelectList).props().children.length;
+    const actual = wrapper.find(SelectList).at(0)
+      .props().children.length;
+
+    expect(actual).equals(expected);
+
+  });
+
+  it('should render 3 list items on the SelectListComponent for category', () => {
+
+    const wrapper = shallow(<ItemDetailForm {...props} />);
+
+    const expected = 3;
+    const actual = wrapper.find(SelectList).at(1)
+      .props().children.length;
 
     expect(actual).equals(expected);
 
@@ -83,9 +138,22 @@ describe('<ItemDetailForm />', () => {
     const wrapper = shallow(<ItemDetailForm {...props} />);
 
     const expected = { name: 'priceTypeID', label: 'Each', value: 0 };
-    const actual = wrapper.find(SelectList).props().value;
+    const actual = wrapper.find(SelectList).at(0)
+      .props().value;
 
     expect(actual).deep.equals(expected);
+
+  });
+
+  it('should contain an "Additional Configuration" heading', () => {
+
+    const wrapper = shallow(<ItemDetailForm {...props} />);
+
+    const expected = 'Additional Configuration';
+    const actual = wrapper.find('.additional-items-heading')
+      .text();
+
+    expect(actual).equals(expected);
 
   });